Add Sphinx to build documentation for DLPack
I will submit a series of PRs to get some documentation up and running for
DLPack! First in the series is a PR to choose a documentation tool and
establish a system to build the docs. In a follow-up PR, I will add a CI
script to automate documentation builds. Then we can start populating the
docs with examples, tutorials, list of projects that use DLPack, etc.

This PR introduces Sphinx as a documentation tool for DLPack. It uses the
Breathe extension to parse Doxygen output (Sphinx doesn't directly parse C/C++
code, so, an extension is required to do that). The new documentation uses
the PyData sphinx theme (adopted by several projects in the data science
community). To keep the diff minimal, I didn't change the default configuration.